# packages/api-contract/openapi-v1.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multicountry Pesticide KG API",
    "version": "1.1"
  },
  "paths": {
...
    "/api/v1/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search",
        "operationId": "search_api_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "jurisdiction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 10
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jurisdiction"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release"
            }
          },
          {
            "name": "X-PestKG-Release",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pestkg-Release"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_EntityData__"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/compare/{question}": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Compare",
        "operationId": "compare_api_v1_compare__question__get",
        "parameters": [
          {
            "name": "question",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "jurisdiction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 10
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jurisdiction"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release"
            }
          },
          {
            "name": "X-PestKG-Release",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pestkg-Release"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_ComparisonRow__"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ComparisonRow": {
        "additionalProperties": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        },
        "type": "object",
        "title": "ComparisonRow"
      },
...
      "EntityData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label_original": {
            "type": "string",
            "title": "Label Original"
          },
          "label_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label En"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "jurisdiction": {
            "type": "string",
            "title": "Jurisdiction"
          },
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "label_original",
          "label_en",
          "type",
          "jurisdiction",
          "source_record_id",
          "source_url",
          "properties"
        ],
        "title": "EntityData"
      },
...
      "Envelope_list_ComparisonRow__": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.1"
          },
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ComparisonRow"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "data"
        ],
        "title": "Envelope[list[ComparisonRow]]"
      },
...
      "Envelope_list_EntityData__": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.1"
          },
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/EntityData"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "data"
        ],
        "title": "Envelope[list[EntityData]]"
      },
...
      "ErrorDescriptor": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDescriptor"
      },
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "title": "Detail"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDescriptor"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
    }
  }
}